pyproject.toml
[tool.poetry]
name = "registrator-romania"
version = "0.1.0"
description = "Script what automaticly registrate users from google sheet on https://cetatenie.just.ro/programare-online/"
authors = ["devtolmachev <dev.tolmachev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
gspread-asyncio = "^2.0.0"
apscheduler = "^3.10.4"
playwright = "^1.44.0"
loguru = "^0.7.2"
envyaml = "^1.10.211231"
pandas = "^2.2.2"
aiogram = "^3.6.0"
bs4 = "^0.0.2"
lxml = "^5.2.2"
fake-useragent = "^1.5.1"
proxybroker = {git = "https://github.com/bluet/proxybroker2.git"}
google-recaptcha = "^2.0.1"
pypasser = "^0.0.5"
ua-generator = "^0.5.0"
esprima = "^4.0.1"
pyjsparser = "^2.7.1"
aiohttp-socks = "^0.8.4"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"
openpyxl = "^3.1.5"
orjson = "^3.10.6"
python-dateutil = "^2.9.0.post0"
requests-toolbelt = "^1.0.0"
slimit = "^0.8.1"
js2py = "^0.74"
python-docx = "^1.1.2"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
click = "^8.1.7"
setuptools = "^71.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
